Consulta antes de eliminar un registro. Lo Agregué en todos los ABM

+++ src/madreteresacrud/GastosABM.java
@@ -13,6 +13,7 @@ import java.util.Calendar;
 import java.util.List;
 import javax.persistence.RollbackException;
 import javax.swing.JFrame;
+import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.SwingUtilities;
 import org.jdesktop.beansbinding.Converter;
@@ -369,17 +370,23 @@ public class GastosABM extends JPanel {
     }//GEN-LAST:event_refreshButtonActionPerformed
     
     private void deleteButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_deleteButtonActionPerformed
-        int[] selected = masterTable.getSelectedRows();
-        List<madreteresacrud.Gastos> toRemove = new ArrayList<madreteresacrud.Gastos>(selected.length);
-        for (int idx = 0; idx < selected.length; idx++) {
-            madreteresacrud.Gastos g = list.get(masterTable.convertRowIndexToModel(selected[idx]));
-            toRemove.add(g);
-            entityManager.remove(g);
+        int reply = JOptionPane.showConfirmDialog(null, "¿Está seguro de eliminar el registro?", "Eliminacion de Registro", JOptionPane.YES_NO_OPTION);
+        if (reply == JOptionPane.YES_OPTION) {
+            JOptionPane.showMessageDialog(null, "Registro Eliminado");
+            int[] selected = masterTable.getSelectedRows();
+            List<madreteresacrud.Gastos> toRemove = new ArrayList<madreteresacrud.Gastos>(selected.length);
+            for (int idx = 0; idx < selected.length; idx++) {
+                madreteresacrud.Gastos g = list.get(masterTable.convertRowIndexToModel(selected[idx]));
+                toRemove.add(g);
+                entityManager.remove(g);
+            }
+            list.removeAll(toRemove);
+
+            jComboBoxElemento.setEnabled(false);
+            jComboBoxTipoGasto.setEnabled(false);  
+            
         }
-        list.removeAll(toRemove);
-        
-        jComboBoxElemento.setEnabled(false);
-        jComboBoxTipoGasto.setEnabled(false);
+      
     }//GEN-LAST:event_deleteButtonActionPerformed
     
     private void newButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_newButtonActionPerformed
