+++ src/madreteresacrud/floresVida/FlorVidaABM.java
@@ -13,6 +13,7 @@ import java.util.Calendar;
 import java.util.List;
 import javax.persistence.RollbackException;
 import javax.swing.JFrame;
+import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.SwingUtilities;
 import org.jdesktop.beansbinding.Converter;
@@ -357,14 +358,18 @@ public class FlorVidaABM extends JPanel {
     }//GEN-LAST:event_refreshButtonActionPerformed
     
     private void deleteButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_deleteButtonActionPerformed
-        int[] selected = masterTable.getSelectedRows();
-        List<madreteresacrud.floresVida.FlorVida> toRemove = new ArrayList<madreteresacrud.floresVida.FlorVida>(selected.length);
-        for (int idx = 0; idx < selected.length; idx++) {
-            madreteresacrud.floresVida.FlorVida f = list.get(masterTable.convertRowIndexToModel(selected[idx]));
-            toRemove.add(f);
-            entityManager.remove(f);
+        int reply = JOptionPane.showConfirmDialog(null, "¿Está seguro de eliminar el registro?", "Eliminacion de Registro", JOptionPane.YES_NO_OPTION);
+        if (reply == JOptionPane.YES_OPTION) {
+            JOptionPane.showMessageDialog(null, "Registro Eliminado");
+            int[] selected = masterTable.getSelectedRows();
+            List<madreteresacrud.floresVida.FlorVida> toRemove = new ArrayList<madreteresacrud.floresVida.FlorVida>(selected.length);
+            for (int idx = 0; idx < selected.length; idx++) {
+                madreteresacrud.floresVida.FlorVida f = list.get(masterTable.convertRowIndexToModel(selected[idx]));
+                toRemove.add(f);
+                entityManager.remove(f);
+            }
+            list.removeAll(toRemove);
         }
-        list.removeAll(toRemove);
     }//GEN-LAST:event_deleteButtonActionPerformed
     
     private void newButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_newButtonActionPerformed
